[Data] Split test_operators.py into smaller, focused test modules
